refactor: 소켓 핸들링 로직 분리
